Restructure folders for uv package setup
