+++ pyproject.toml
@@ -22,8 +22,5 @@ dependencies = [
     "qdrant-client>=1.14.1",
 ]
 
-[tool.setuptools]
-packages = ["rag-who"]
-
 [tool.uv]
 package = true
